requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-crispy-forms==1.14.0
django-decouple==2.1
idna==3.4
psycopg2==2.9.5
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
joblib==1.2.0
numpy==1.23.5
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.6
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
xgboost==1.7.1
click==8.1.3
Flask==2.2.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.11.0
waitress==2.1.2
kafka-python==2.0.2